+++ lib/templates/getAllTemplates.ts
@@ -2,7 +2,7 @@ import { prisma, prismaErrors, Prisma } from "@gcforms/database";
 import { FormRecord } from "@lib/types";
 import { authorization } from "../privileges";
 import { AuditLogAccessDeniedDetails, AuditLogDetails, logEvent } from "../auditLogs";
-import { parseTemplate } from "./shared";
+import { parseTemplate } from "./internal";
 import { logMessage } from "@lib/logger";
 
 /**
