+++ lib/templates/getAllTemplatesForUser.ts
@@ -3,7 +3,7 @@ import { FormRecord } from "@lib/types";
 import { getAbility } from "../privileges";
 import { AuditLogDetails, logEvent } from "../auditLogs";
 import { logMessage } from "@lib/logger";
-import { parseTemplate } from "./shared";
+import { parseTemplate } from "./internal";
 
 export type TemplateOptions = {
   sortByDateUpdated?: "asc" | "desc";
