+++ lib/templates/getFullTemplateByID.ts
@@ -2,7 +2,7 @@ import { FormRecord } from "@lib/types";
 import { prisma, prismaErrors } from "@gcforms/database";
 import { authorization } from "../privileges";
 import { AuditLogAccessDeniedDetails, logEvent } from "../auditLogs";
-import { parseTemplate } from "./shared";
+import { parseTemplate } from "./internal";
 
 /**
  * Get a form template by ID (includes full template information but requires view permission)
