+++ lib/templates/internal/index.ts
@@ -1,5 +1,5 @@
 import { Prisma } from "@gcforms/database";
-import { authorization } from "../privileges";
+import { authorization } from "../../privileges";
 import { checkOne } from "@lib/cache/flags";
 import { FormRecord, FormProperties, SecurityAttribute, ClosedDetails } from "@lib/types";
 import { NotificationsInterval, PublicFormRecord } from "@gcforms/types";
