+++ lib/templates/updateFormPurpose.ts
@@ -4,7 +4,7 @@ import { FormRecord } from "@lib/types";
 import { authorization } from "../privileges";
 import { AuditLogAccessDeniedDetails, AuditLogDetails, logEvent } from "../auditLogs";
 import { TemplateAlreadyPublishedError } from "./internal/errors";
-import { parseTemplate } from "./shared";
+import { parseTemplate } from "./internal";
 
 export async function updateFormPurpose(
   formID: string,
